+++ src/configs/verification-challenges.js
@@ -5,6 +5,7 @@ const {
   USERS_ACTION_DISPOSABLE_PASSWORD,
   USERS_ACTION_REGISTER,
   USERS_ACTION_RESET,
+  USERS_ACTION_VERIFY_CONTACT,
 } = require('../constants');
 
 /**
@@ -58,10 +59,19 @@ exports.phone = {
   account: 'replace-with-your-account',
   messages: {
     [USERS_ACTION_ACTIVATE]: '%s is your activation code',
+    [USERS_ACTION_VERIFY_CONTACT]: '%s is your verification code',
     [USERS_ACTION_DISPOSABLE_PASSWORD]: '%s is your disposable password',
     [USERS_ACTION_REGISTER]: '%s is your password',
     [USERS_ACTION_RESET]: '%s is your code for reset password',
   },
   prefix: 'phone',
   waitChallenge: false,
+};
+
+/**
+ * User contacts settings
+ * @type {Object}
+ */
+exports.contacts = {
+  max: 5,
 };