+++ test/suites/actions/oauth/facebook.js
@@ -213,7 +213,6 @@ describe('#facebook', function oauthFacebookSuite() {
       });
     });
 
-
     /**
      * Internal check. Just to be sure if Throttling Happened during test suites.
      */
@@ -269,8 +268,9 @@ describe('#facebook', function oauthFacebookSuite() {
    * Application has any access to the users Facebook profile.
    * This suite don't need to recreate user for each test and we can use one AuthToken in all tests.
    */
-  describe('new user', async () => {
+  describe('new user', async function newUserTest() {
     let generalUser;
+    this.slow(60000);
 
     before('create test user', async () => {
       generalUser = await GraphApi.createTestUser();
@@ -446,7 +446,6 @@ describe('#facebook', function oauthFacebookSuite() {
         await globalRegisterUser(username).call(dataBag);
         await globalAuthUser(username).call(dataBag);
       });
-
 
       after('stop executer', async () => {
         await fb.stop();